+++ client/src/routeTree.gen.ts
@@ -24,13 +24,12 @@ import { Route as GameidChatsRouteImport } from './routes/$gameid/chats'
 import { Route as GameidCompareRouteImport } from './routes/$gameid/compare'
 import { Route as GameidDriverRouteImport } from './routes/$gameid/driver'
 import { Route as GameidExperimentsRouteImport } from './routes/$gameid/experiments'
+import { Route as GameidRawRouteImport } from './routes/$gameid/raw'
 import { Route as GameidSessionsRouteImport } from './routes/$gameid/sessions'
 import { Route as GameidTracksRouteImport } from './routes/$gameid/tracks'
 import { Route as AcEvoIndexRouteImport } from './routes/ac-evo/index'
-import { Route as AcEvoRawRouteImport } from './routes/ac-evo/raw'
 import { Route as AcEvoSetupsRouteImport } from './routes/ac-evo/setups'
 import { Route as AccIndexRouteImport } from './routes/acc/index'
-import { Route as AccRawRouteImport } from './routes/acc/raw'
 import { Route as AccSetupsRouteImport } from './routes/acc/setups'
 import { Route as DashIndexRouteImport } from './routes/dash.index'
 import { Route as DashCombo1RouteImport } from './routes/dash.combo-1'
@@ -38,12 +37,10 @@ import { Route as DashCombo2RouteImport } from './routes/dash.combo-2'
 import { Route as DevIndexRouteImport } from './routes/dev.index'
 import { Route as DevTracksRouteImport } from './routes/dev.tracks'
 import { Route as F125IndexRouteImport } from './routes/f125/index'
-import { Route as F125RawRouteImport } from './routes/f125/raw'
 import { Route as F125SetupsRouteImport } from './routes/f125/setups'
 import { Route as F125TunesRouteImport } from './routes/f125/tunes'
 import { Route as Fm23IndexRouteImport } from './routes/fm23/index'
 import { Route as Fm23LiveRouteImport } from './routes/fm23/live'
-import { Route as Fm23RawRouteImport } from './routes/fm23/raw'
 import { Route as Fm23SetupsRouteImport } from './routes/fm23/setups'
 import { Route as IracingIndexRouteImport } from './routes/iracing/index'
 import { Route as IracingLiveRouteImport } from './routes/iracing/live'
@@ -154,6 +151,11 @@ const GameidExperimentsRoute = GameidExperimentsRouteImport.update({
   path: '/experiments',
   getParentRoute: () => GameidRoute,
 } as any)
+const GameidRawRoute = GameidRawRouteImport.update({
+  id: '/raw',
+  path: '/raw',
+  getParentRoute: () => GameidRoute,
+} as any)
 const GameidSessionsRoute = GameidSessionsRouteImport.update({
   id: '/sessions',
   path: '/sessions',
@@ -169,11 +171,6 @@ const AcEvoIndexRoute = AcEvoIndexRouteImport.update({
   path: '/',
   getParentRoute: () => AcEvoRoute,
 } as any)
-const AcEvoRawRoute = AcEvoRawRouteImport.update({
-  id: '/raw',
-  path: '/raw',
-  getParentRoute: () => AcEvoRoute,
-} as any)
 const AcEvoSetupsRoute = AcEvoSetupsRouteImport.update({
   id: '/setups',
   path: '/setups',
@@ -182,11 +179,6 @@ const AcEvoSetupsRoute = AcEvoSetupsRouteImport.update({
 const AccIndexRoute = AccIndexRouteImport.update({
   id: '/',
   path: '/',
-  getParentRoute: () => AccRoute,
-} as any)
-const AccRawRoute = AccRawRouteImport.update({
-  id: '/raw',
-  path: '/raw',
   getParentRoute: () => AccRoute,
 } as any)
 const AccSetupsRoute = AccSetupsRouteImport.update({
@@ -224,11 +216,6 @@ const F125IndexRoute = F125IndexRouteImport.update({
   path: '/',
   getParentRoute: () => F125Route,
 } as any)
-const F125RawRoute = F125RawRouteImport.update({
-  id: '/raw',
-  path: '/raw',
-  getParentRoute: () => F125Route,
-} as any)
 const F125SetupsRoute = F125SetupsRouteImport.update({
   id: '/setups',
   path: '/setups',
@@ -247,11 +234,6 @@ const Fm23IndexRoute = Fm23IndexRouteImport.update({
 const Fm23LiveRoute = Fm23LiveRouteImport.update({
   id: '/live',
   path: '/live',
-  getParentRoute: () => Fm23Route,
-} as any)
-const Fm23RawRoute = Fm23RawRouteImport.update({
-  id: '/raw',
-  path: '/raw',
   getParentRoute: () => Fm23Route,
 } as any)
 const Fm23SetupsRoute = Fm23SetupsRouteImport.update({
@@ -450,20 +432,17 @@ export interface FileRoutesByFullPath {
   '/$gameid/compare': typeof GameidCompareRoute
   '/$gameid/driver': typeof GameidDriverRoute
   '/$gameid/experiments': typeof GameidExperimentsRouteWithChildren
+  '/$gameid/raw': typeof GameidRawRoute
   '/$gameid/sessions': typeof GameidSessionsRoute
   '/$gameid/tracks': typeof GameidTracksRouteWithChildren
-  '/ac-evo/raw': typeof AcEvoRawRoute
   '/ac-evo/setups': typeof AcEvoSetupsRouteWithChildren
-  '/acc/raw': typeof AccRawRoute
   '/acc/setups': typeof AccSetupsRouteWithChildren
   '/dash/combo-1': typeof DashCombo1Route
   '/dash/combo-2': typeof DashCombo2Route
   '/dev/tracks': typeof DevTracksRouteWithChildren
-  '/f125/raw': typeof F125RawRoute
   '/f125/setups': typeof F125SetupsRouteWithChildren
   '/f125/tunes': typeof F125TunesRouteWithChildren
   '/fm23/live': typeof Fm23LiveRouteWithChildren
-  '/fm23/raw': typeof Fm23RawRoute
   '/fm23/setups': typeof Fm23SetupsRouteWithChildren
   '/iracing/live': typeof IracingLiveRouteWithChildren
   '/ac-evo/': typeof AcEvoIndexRoute
@@ -514,14 +493,11 @@ export interface FileRoutesByTo {
   '/$gameid/chats': typeof GameidChatsRoute
   '/$gameid/compare': typeof GameidCompareRoute
   '/$gameid/driver': typeof GameidDriverRoute
+  '/$gameid/raw': typeof GameidRawRoute
   '/$gameid/sessions': typeof GameidSessionsRoute
-  '/ac-evo/raw': typeof AcEvoRawRoute
-  '/acc/raw': typeof AccRawRoute
   '/dash/combo-1': typeof DashCombo1Route
   '/dash/combo-2': typeof DashCombo2Route
-  '/f125/raw': typeof F125RawRoute
   '/fm23/live': typeof Fm23LiveRouteWithChildren
-  '/fm23/raw': typeof Fm23RawRoute
   '/iracing/live': typeof IracingLiveRouteWithChildren
   '/ac-evo': typeof AcEvoIndexRoute
   '/acc': typeof AccIndexRoute
@@ -578,20 +554,17 @@ export interface FileRoutesById {
   '/$gameid/compare': typeof GameidCompareRoute
   '/$gameid/driver': typeof GameidDriverRoute
   '/$gameid/experiments': typeof GameidExperimentsRouteWithChildren
+  '/$gameid/raw': typeof GameidRawRoute
   '/$gameid/sessions': typeof GameidSessionsRoute
   '/$gameid/tracks': typeof GameidTracksRouteWithChildren
-  '/ac-evo/raw': typeof AcEvoRawRoute
   '/ac-evo/setups': typeof AcEvoSetupsRouteWithChildren
-  '/acc/raw': typeof AccRawRoute
   '/acc/setups': typeof AccSetupsRouteWithChildren
   '/dash/combo-1': typeof DashCombo1Route
   '/dash/combo-2': typeof DashCombo2Route
   '/dev/tracks': typeof DevTracksRouteWithChildren
-  '/f125/raw': typeof F125RawRoute
   '/f125/setups': typeof F125SetupsRouteWithChildren
   '/f125/tunes': typeof F125TunesRouteWithChildren
   '/fm23/live': typeof Fm23LiveRouteWithChildren
-  '/fm23/raw': typeof Fm23RawRoute
   '/fm23/setups': typeof Fm23SetupsRouteWithChildren
   '/iracing/live': typeof IracingLiveRouteWithChildren
   '/ac-evo/': typeof AcEvoIndexRoute
@@ -651,20 +624,17 @@ export interface FileRouteTypes {
     | '/$gameid/compare'
     | '/$gameid/driver'
     | '/$gameid/experiments'
+    | '/$gameid/raw'
     | '/$gameid/sessions'
     | '/$gameid/tracks'
-    | '/ac-evo/raw'
     | '/ac-evo/setups'
-    | '/acc/raw'
     | '/acc/setups'
     | '/dash/combo-1'
     | '/dash/combo-2'
     | '/dev/tracks'
-    | '/f125/raw'
     | '/f125/setups'
     | '/f125/tunes'
     | '/fm23/live'
-    | '/fm23/raw'
     | '/fm23/setups'
     | '/iracing/live'
     | '/ac-evo/'
@@ -715,14 +685,11 @@ export interface FileRouteTypes {
     | '/$gameid/chats'
     | '/$gameid/compare'
     | '/$gameid/driver'
+    | '/$gameid/raw'
     | '/$gameid/sessions'
-    | '/ac-evo/raw'
-    | '/acc/raw'
     | '/dash/combo-1'
     | '/dash/combo-2'
-    | '/f125/raw'
     | '/fm23/live'
-    | '/fm23/raw'
     | '/iracing/live'
     | '/ac-evo'
     | '/acc'
@@ -778,20 +745,17 @@ export interface FileRouteTypes {
     | '/$gameid/compare'
     | '/$gameid/driver'
     | '/$gameid/experiments'
+    | '/$gameid/raw'
     | '/$gameid/sessions'
     | '/$gameid/tracks'
-    | '/ac-evo/raw'
     | '/ac-evo/setups'
-    | '/acc/raw'
     | '/acc/setups'
     | '/dash/combo-1'
     | '/dash/combo-2'
     | '/dev/tracks'
-    | '/f125/raw'
     | '/f125/setups'
     | '/f125/tunes'
     | '/fm23/live'
-    | '/fm23/raw'
     | '/fm23/setups'
     | '/iracing/live'
     | '/ac-evo/'
@@ -956,6 +920,13 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof GameidExperimentsRouteImport
       parentRoute: typeof GameidRoute
     }
+    '/$gameid/raw': {
+      id: '/$gameid/raw'
+      path: '/raw'
+      fullPath: '/$gameid/raw'
+      preLoaderRoute: typeof GameidRawRouteImport
+      parentRoute: typeof GameidRoute
+    }
     '/$gameid/sessions': {
       id: '/$gameid/sessions'
       path: '/sessions'
@@ -977,13 +948,6 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof AcEvoIndexRouteImport
       parentRoute: typeof AcEvoRoute
     }
-    '/ac-evo/raw': {
-      id: '/ac-evo/raw'
-      path: '/raw'
-      fullPath: '/ac-evo/raw'
-      preLoaderRoute: typeof AcEvoRawRouteImport
-      parentRoute: typeof AcEvoRoute
-    }
     '/ac-evo/setups': {
       id: '/ac-evo/setups'
       path: '/setups'
@@ -996,13 +960,6 @@ declare module '@tanstack/react-router' {
       path: '/'
       fullPath: '/acc/'
       preLoaderRoute: typeof AccIndexRouteImport
-      parentRoute: typeof AccRoute
-    }
-    '/acc/raw': {
-      id: '/acc/raw'
-      path: '/raw'
-      fullPath: '/acc/raw'
-      preLoaderRoute: typeof AccRawRouteImport
       parentRoute: typeof AccRoute
     }
     '/acc/setups': {
@@ -1054,13 +1011,6 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof F125IndexRouteImport
       parentRoute: typeof F125Route
     }
-    '/f125/raw': {
-      id: '/f125/raw'
-      path: '/raw'
-      fullPath: '/f125/raw'
-      preLoaderRoute: typeof F125RawRouteImport
-      parentRoute: typeof F125Route
-    }
     '/f125/setups': {
       id: '/f125/setups'
       path: '/setups'
@@ -1087,13 +1037,6 @@ declare module '@tanstack/react-router' {
       path: '/live'
       fullPath: '/fm23/live'
       preLoaderRoute: typeof Fm23LiveRouteImport
-      parentRoute: typeof Fm23Route
-    }
-    '/fm23/raw': {
-      id: '/fm23/raw'
-      path: '/raw'
-      fullPath: '/fm23/raw'
-      preLoaderRoute: typeof Fm23RawRouteImport
       parentRoute: typeof Fm23Route
     }
     '/fm23/setups': {
@@ -1376,6 +1319,7 @@ interface GameidRouteChildren {
   GameidCompareRoute: typeof GameidCompareRoute
   GameidDriverRoute: typeof GameidDriverRoute
   GameidExperimentsRoute: typeof GameidExperimentsRouteWithChildren
+  GameidRawRoute: typeof GameidRawRoute
   GameidSessionsRoute: typeof GameidSessionsRoute
   GameidTracksRoute: typeof GameidTracksRouteWithChildren
 }
@@ -1387,6 +1331,7 @@ const GameidRouteChildren: GameidRouteChildren = {
   GameidCompareRoute: GameidCompareRoute,
   GameidDriverRoute: GameidDriverRoute,
   GameidExperimentsRoute: GameidExperimentsRouteWithChildren,
+  GameidRawRoute: GameidRawRoute,
   GameidSessionsRoute: GameidSessionsRoute,
   GameidTracksRoute: GameidTracksRouteWithChildren,
 }
@@ -1413,13 +1358,11 @@ const AcEvoSetupsRouteWithChildren = AcEvoSetupsRoute._addFileChildren(
 )
 
 interface AcEvoRouteChildren {
-  AcEvoRawRoute: typeof AcEvoRawRoute
   AcEvoSetupsRoute: typeof AcEvoSetupsRouteWithChildren
   AcEvoIndexRoute: typeof AcEvoIndexRoute
 }
 
 const AcEvoRouteChildren: AcEvoRouteChildren = {
-  AcEvoRawRoute: AcEvoRawRoute,
   AcEvoSetupsRoute: AcEvoSetupsRouteWithChildren,
   AcEvoIndexRoute: AcEvoIndexRoute,
 }
@@ -1445,13 +1388,11 @@ const AccSetupsRouteWithChildren = AccSetupsRoute._addFileChildren(
 )
 
 interface AccRouteChildren {
-  AccRawRoute: typeof AccRawRoute
   AccSetupsRoute: typeof AccSetupsRouteWithChildren
   AccIndexRoute: typeof AccIndexRoute
 }
 
 const AccRouteChildren: AccRouteChildren = {
-  AccRawRoute: AccRawRoute,
   AccSetupsRoute: AccSetupsRouteWithChildren,
   AccIndexRoute: AccIndexRoute,
 }
@@ -1534,14 +1475,12 @@ const F125TunesRouteWithChildren = F125TunesRoute._addFileChildren(
 )
 
 interface F125RouteChildren {
-  F125RawRoute: typeof F125RawRoute
   F125SetupsRoute: typeof F125SetupsRouteWithChildren
   F125TunesRoute: typeof F125TunesRouteWithChildren
   F125IndexRoute: typeof F125IndexRoute
 }
 
 const F125RouteChildren: F125RouteChildren = {
-  F125RawRoute: F125RawRoute,
   F125SetupsRoute: F125SetupsRouteWithChildren,
   F125TunesRoute: F125TunesRouteWithChildren,
   F125IndexRoute: F125IndexRoute,
@@ -1583,7 +1522,6 @@ const Fm23SetupsRouteWithChildren = Fm23SetupsRoute._addFileChildren(
 
 interface Fm23RouteChildren {
   Fm23LiveRoute: typeof Fm23LiveRouteWithChildren
-  Fm23RawRoute: typeof Fm23RawRoute
   Fm23SetupsRoute: typeof Fm23SetupsRouteWithChildren
   Fm23IndexRoute: typeof Fm23IndexRoute
   Fm23CarsCarOrdinalRoute: typeof Fm23CarsCarOrdinalRoute
@@ -1591,7 +1529,6 @@ interface Fm23RouteChildren {
 
 const Fm23RouteChildren: Fm23RouteChildren = {
   Fm23LiveRoute: Fm23LiveRouteWithChildren,
-  Fm23RawRoute: Fm23RawRoute,
   Fm23SetupsRoute: Fm23SetupsRouteWithChildren,
   Fm23IndexRoute: Fm23IndexRoute,
   Fm23CarsCarOrdinalRoute: Fm23CarsCarOrdinalRoute,